Backend and flask improvements
Fixing up the backend ``aiohttp.web`` appliction so that the backend Cog will be fully responsible for managing the lifecycle but views / handlers will be imported much like blueprints are.

Finally, trying to get to the bottom of some of the quirks of getting either the Flask application or Discord bot to run under a seperate thread. Neither seems to play well with a non-main thread event loop since both make use of signals that are not available on non-main threads..

So far it seems "easier" to run Flask in a seperate thread but I am also considering a completely new process since the bot and Flask application have no direct connection and only communicate through the API/RPC interfaces.

Also adding "build" meta requirement in pyproject.toml